chore: update entrypoint
