+++ src/create_pytauri_app/questionnaire.py
@@ -1,4 +1,3 @@
-from dataclasses import asdict, dataclass
 from pathlib import Path
 from typing import Literal
